# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QuantPilot Operator Pre-Harness",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/portfolio/plan": {
      "post": {
        "summary": "Create Portfolio Plan",
        "operationId": "create_portfolio_plan_api_portfolio_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioPlan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/plan": {
      "post": {
        "summary": "Create Order Plans",
        "operationId": "create_order_plans_api_orders_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrderPlan"
                  },
                  "type": "array",
                  "title": "Response Create Order Plans Api Orders Plan Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/generate-proposals": {
      "post": {
        "summary": "Generate Order Proposals",
        "operationId": "generate_order_proposals_api_orders_generate_proposals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrderPlan"
                  },
                  "type": "array",
                  "title": "Response Generate Order Proposals Api Orders Generate Proposals Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orders/{order_plan_id}/approve": {
      "post": {
        "summary": "Approve Order",
        "operationId": "approve_order_api_orders__order_plan_id__approve_post",
        "parameters": [
          {
            "name": "order_plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderPlan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{order_plan_id}/reject": {
      "post": {
        "summary": "Reject Order",
        "operationId": "reject_order_api_orders__order_plan_id__reject_post",
        "parameters": [
          {
            "name": "order_plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectOrderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderPlan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{order_plan_id}/modify": {
      "post": {
        "summary": "Modify Order",
        "operationId": "modify_order_api_orders__order_plan_id__modify_post",
        "parameters": [
          {
            "name": "order_plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyOrderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderPlan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{order_plan_id}/submit": {
      "post": {
        "summary": "Submit Order",
        "operationId": "submit_order_api_orders__order_plan_id__submit_post",
        "parameters": [
          {
            "name": "order_plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Submit Order Api Orders  Order Plan Id  Submit Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{order_plan_id}/status": {
      "get": {
        "summary": "Order Status",
        "operationId": "order_status_api_orders__order_plan_id__status_get",
        "parameters": [
          {
            "name": "order_plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Order Status Api Orders  Order Plan Id  Status Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModifyOrderRequest": {
        "properties": {
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "limit_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Price"
          }
        },
        "type": "object",
        "required": [
          "quantity"
        ],
        "title": "ModifyOrderRequest"
      },
...
      "OrderIntent": {
        "properties": {
          "intent_id": {
            "type": "string",
            "title": "Intent Id"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "side": {
            "type": "string",
            "enum": [
              "buy",
              "sell"
            ],
            "title": "Side"
          },
          "order_type": {
            "$ref": "#/components/schemas/OrderType",
            "default": "limit"
          },
          "quantity": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Quantity"
          },
          "limit_price": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Price"
          },
          "notional": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Notional"
          },
          "target_weight": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Target Weight"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "quote_time": {
            "type": "string",
            "format": "date-time",
            "title": "Quote Time"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "symbol",
          "side",
          "quantity",
          "notional",
          "target_weight",
          "reason"
        ],
        "title": "OrderIntent"
      },
      "OrderPlan": {
        "properties": {
          "order_plan_id": {
            "type": "string",
            "title": "Order Plan Id"
          },
          "policy_id": {
            "type": "string",
            "title": "Policy Id"
          },
          "policy_version": {
            "type": "integer",
            "title": "Policy Version"
          },
          "intent": {
            "$ref": "#/components/schemas/OrderIntent"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus",
            "default": "draft"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "risk_check_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Check Id"
          },
          "risk_check_expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Check Expires At"
          },
          "explanation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProposalExplanation"
              },
              {
                "type": "null"
              }
            ]
          },
          "auto_order_reference_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Order Reference Price"
          },
          "replaces_order_plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replaces Order Plan Id"
          },
          "blocked_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocked Reason"
          },
          "approved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "policy_id",
          "policy_version",
          "intent",
          "idempotency_key"
        ],
        "title": "OrderPlan"
      },
      "OrderPlanRequest": {
        "properties": {
          "portfolio_plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portfolio Plan Id"
          }
        },
        "type": "object",
        "title": "OrderPlanRequest"
      },
      "OrderStatus": {
        "type": "string",
        "enum": [
          "draft",
          "risk_checked",
          "proposed",
          "modified",
          "user_approved",
          "submitted",
          "accepted",
          "partially_filled",
          "filled",
          "cancelled",
          "rejected",
          "expired",
          "failed"
        ],
        "title": "OrderStatus"
      },
      "OrderType": {
        "type": "string",
        "enum": [
          "limit",
          "market"
        ],
        "title": "OrderType"
      },
...
      "PortfolioPlan": {
        "properties": {
          "plan_id": {
            "type": "string",
            "title": "Plan Id"
          },
          "policy_id": {
            "type": "string",
            "title": "Policy Id"
          },
          "policy_version": {
            "type": "integer",
            "title": "Policy Version"
          },
          "target_weights": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Target Weights"
          },
          "cash_target_weight": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Cash Target Weight"
          },
          "order_intents": {
            "items": {
              "$ref": "#/components/schemas/OrderIntent"
            },
            "type": "array",
            "title": "Order Intents"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "policy_id",
          "policy_version",
          "target_weights",
          "cash_target_weight",
          "order_intents"
        ],
        "title": "PortfolioPlan"
      },
      "PortfolioPlanRequest": {
        "properties": {
          "policy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Id"
          }
        },
        "type": "object",
        "title": "PortfolioPlanRequest"
      },
      "ProposalExplanation": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "action": {
            "type": "string",
            "enum": [
              "buy",
              "sell"
            ],
            "title": "Action"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "target_weight_delta": {
            "type": "number",
            "title": "Target Weight Delta"
          },
          "reference_price": {
            "type": "number",
            "title": "Reference Price"
          },
          "estimated_cash_impact": {
            "type": "number",
            "title": "Estimated Cash Impact"
          },
          "strategy_id": {
            "type": "string",
            "title": "Strategy Id"
          },
          "strategy_version": {
            "type": "string",
            "title": "Strategy Version"
          },
          "signal_reason": {
            "type": "string",
            "title": "Signal Reason"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reason Codes"
          },
          "current_weight": {
            "type": "number",
            "title": "Current Weight"
          },
          "target_weight": {
            "type": "number",
            "title": "Target Weight"
          },
          "weight_delta": {
            "type": "number",
            "title": "Weight Delta"
          },
          "quote_price": {
            "type": "number",
            "title": "Quote Price"
          },
          "quote_age_seconds": {
            "type": "number",
            "title": "Quote Age Seconds"
          },
          "limit_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Price"
          },
          "estimated_notional": {
            "type": "number",
            "title": "Estimated Notional"
          },
          "estimated_cost_bps": {
            "type": "number",
            "title": "Estimated Cost Bps",
            "default": 0.0
          },
          "stop_price_hint": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Price Hint"
          },
          "take_profit_hint": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Take Profit Hint"
          },
          "risk_checks_passed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Risk Checks Passed"
          },
          "risk_checks_failed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Risk Checks Failed"
          },
          "risk_check_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Check Id"
          },
          "risk_check_expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Check Expires At"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "policy_version": {
            "type": "integer",
            "title": "Policy Version"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "symbol",
          "action",
          "quantity",
          "target_weight_delta",
          "reference_price",
          "estimated_cash_impact",
          "strategy_id",
          "strategy_version",
          "signal_reason",
          "current_weight",
          "target_weight",
          "weight_delta",
          "quote_price",
          "quote_age_seconds",
          "estimated_notional",
          "idempotency_key",
          "policy_version"
        ],
        "title": "ProposalExplanation"
      },
      "RejectOrderRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": "user_rejected"
          }
        },
        "type": "object",
        "title": "RejectOrderRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}